Refaktor. Trekk dataklasser ut til egen fil for ryddigere oversikt
